* Fixed: More Signed-Unsigned comparison warnings
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4033 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/codecompletion/ccoptionsprjdlg.cpp
@@ -122,7 +122,7 @@ void CCOptionsProjectDlg::OnApply()
 {
     wxArrayString newpaths;
     wxListBox* control = XRCCTRL(*this, "lstPaths", wxListBox);
-    for (int i = 0; i < control->GetCount(); ++i)
+    for (int i = 0; i < (int)control->GetCount(); ++i)
     {
         newpaths.Add(control->GetString(i));
     }
